+++ src/System/MemInfo.hs
@@ -56,8 +56,10 @@ module System.MemInfo (
 
 import Data.Bifunctor (Bifunctor (..))
 import Data.Functor ((<&>))
+import Data.List (sortBy)
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
+import Data.Ord (Down (..), comparing)
 import qualified Data.Text as Text
 import qualified Data.Text.IO as Text
 import Fmt (
@@ -67,7 +69,7 @@ import Fmt (
   (|++|),
  )
 import System.Exit (exitFailure)
-import System.MemInfo.Choices (Choices (..), getChoices)
+import System.MemInfo.Choices (Choices (..), PrintOrder (..), getChoices)
 import System.MemInfo.Prelude
 import System.MemInfo.Print (
   AsCmdName (..),
@@ -100,53 +102,64 @@ import System.Posix.User (getEffectiveUserID)
 specified by @Choices@
 -}
 printProcs :: Choices -> IO ()
-printProcs cs = do
+printProcs cs@Choices {choiceByPid = byPid} = do
   bud <- verify cs
+  if byPid
+    then printProcs' withPid bud cs
+    else printProcs' dropId bud cs
+
+
+printProcs' :: (Ord a, AsCmdName a) => Indexer a -> ReportBud -> Choices -> IO ()
+printProcs' indexer bud cs = do
   let Choices
         { choiceShowSwap = showSwap
         , choiceOnlyTotal = onlyTotal
         , choiceWatchSecs = watchSecsMb
-        , choiceByPid = byPid
+        , choicePrintOrder = printOrder
+        , choiceReversed = reversed
         } = cs
-      printEachCmd totals = printMemUsages bud showSwap onlyTotal totals
-      printTheTotal = onlyPrintTotal bud showSwap onlyTotal
-      showTotal cmds = if onlyTotal then printTheTotal cmds else printEachCmd cmds
+      toList = sortBy (byPrintOrder' reversed printOrder) . Map.toList
+      printEachCmd = printMemUsages bud showSwap onlyTotal . toList
+      printTheTotal = onlyPrintTotal bud showSwap onlyTotal . toList
+      showTotal = if onlyTotal then printTheTotal else printEachCmd
       namer = if choiceSplitArgs cs then nameAsFullCmd else nameFor
-  case (watchSecsMb, byPid) of
-    (Nothing, True) -> readMemUsage' namer withPid bud >>= either haltLostPid showTotal
-    (Nothing, _) -> readMemUsage' namer dropId bud >>= either haltLostPid showTotal
-    (Just spanSecs, True) -> do
-      let unfold = unfoldMemUsageAfter' namer withPid spanSecs
-      loopPrintMemUsages unfold bud showTotal
-    (Just spanSecs, _) -> do
-      let unfold = unfoldMemUsageAfter' namer dropId spanSecs
+  case (watchSecsMb) of
+    Nothing -> readMemUsage' namer indexer bud >>= either haltLostPid showTotal
+    (Just spanSecs) -> do
+      let unfold = unfoldMemUsageAfter' namer indexer spanSecs
       loopPrintMemUsages unfold bud showTotal
 
 
-printMemUsages :: AsCmdName a => ReportBud -> Bool -> Bool -> Map a MemUsage -> IO ()
+printMemUsages ::
+  (AsCmdName a) =>
+  ReportBud ->
+  Bool ->
+  Bool ->
+  [(a, MemUsage)] ->
+  IO ()
 printMemUsages bud showSwap onlyTotal totals = do
-  let overall = overallTotals $ Map.elems totals
+  let overall = overallTotals $ map snd totals
       overallIsAccurate = (showSwap && rbHasSwapPss bud) || rbHasPss bud
       print' (name, stats) = Text.putStrLn $ fmtMemUsage showSwap name stats
   Text.putStrLn $ fmtAsHeader showSwap
-  mapM_ print' $ Map.toList totals
+  mapM_ print' totals
   when overallIsAccurate $ Text.putStrLn $ fmtOverall showSwap overall
   reportFlaws bud showSwap onlyTotal
 
 
 -- | Print the program name and memory usage, optionally hiding the swap size
-printUsage' :: AsCmdName a => (a, MemUsage) -> Bool -> IO ()
+printUsage' :: (AsCmdName a) => (a, MemUsage) -> Bool -> IO ()
 printUsage' (name, mu) showSwap = Text.putStrLn $ fmtMemUsage showSwap name mu
 
 
 -- | Like @'printUsage''@, but alway shows the swap size
-printUsage :: AsCmdName a => (a, MemUsage) -> IO ()
+printUsage :: (AsCmdName a) => (a, MemUsage) -> IO ()
 printUsage = flip printUsage' True
 
 
-onlyPrintTotal :: ReportBud -> Bool -> Bool -> Map k MemUsage -> IO ()
+onlyPrintTotal :: (AsCmdName k) => ReportBud -> Bool -> Bool -> [(k, MemUsage)] -> IO ()
 onlyPrintTotal bud showSwap onlyTotal totals = do
-  let (private, swap) = overallTotals $ Map.elems totals
+  let (private, swap) = overallTotals $ map snd totals
       printRawTotal = Text.putStrLn . fmtMemBytes
   if showSwap
     then do
@@ -190,18 +203,18 @@ warnStopped pids = unless (null pids) $ do
 type ProcName = Text
 
 
--- | Like @'unfoldMemUsageAfter''@, but uses the default 'ProcName' and 'Indexer'
+-- | Like @'unfoldMemUsageAfter''@, but uses the default 'ProcNamer' and 'Indexer'
 unfoldMemUsageAfter ::
   (Integral seconds) =>
   seconds ->
   ReportBud ->
-  IO (Either [ProcessID] ((Map Text MemUsage, [ProcessID]), ReportBud))
+  IO (Either [ProcessID] ((Map ProcName MemUsage, [ProcessID]), ReportBud))
 unfoldMemUsageAfter = unfoldMemUsageAfter' nameFor dropId
 
 
 -- | Like @'unfoldMemUsage'@ but computes the @'MemUsage's@ after a delay
 unfoldMemUsageAfter' ::
-  (Ord a, Integral seconds) =>
+  (Ord a, AsCmdName a, Integral seconds) =>
   ProcNamer ->
   Indexer a ->
   seconds ->
@@ -266,7 +279,7 @@ Fails if
 - any of the processes specified by @'ReportBud'@ are missing or inaccessible
 -}
 readMemUsage' ::
-  Ord a =>
+  (Ord a) =>
   ProcNamer ->
   Indexer a ->
   ReportBud ->
@@ -487,7 +500,9 @@ allKnownProcs =
       orNoPids = maybe (Left NoRecords) Right
    in readNaturals (listDirectory procRoot)
         >>= filterM pidExeExists
-        >>= pure . orNoPids . nonEmpty
+        >>= pure
+        . orNoPids
+        . nonEmpty
 
 
 baseName :: Text -> Text
@@ -498,12 +513,12 @@ readMemStats :: ReportBud -> ProcessID -> IO (Either LostPid ProcUsage)
 readMemStats bud pid = do
   statmExists <- doesFileExist $ pidPath "statm" pid
   if
-      | rbHasSmaps bud -> Right . parseFromSmap <$> readSmaps pid
-      | statmExists -> do
-          let readStatm' = readUtf8Text $ pidPath "statm" pid
-              orLostPid = maybe (Left $ BadStatm pid) Right
-          orLostPid . parseFromStatm (rbKernel bud) <$> readStatm'
-      | otherwise -> pure $ Left $ NoProc pid
+    | rbHasSmaps bud -> Right . parseFromSmap <$> readSmaps pid
+    | statmExists -> do
+        let readStatm' = readUtf8Text $ pidPath "statm" pid
+            orLostPid = maybe (Left $ BadStatm pid) Right
+        orLostPid . parseFromStatm (rbKernel bud) <$> readStatm'
+    | otherwise -> pure $ Left $ NoProc pid
 
 
 readSmaps :: ProcessID -> IO Text
@@ -513,9 +528,9 @@ readSmaps pid = do
   hasSmaps <- doesFileExist smapPath
   hasRollup <- doesFileExist rollupPath
   if
-      | hasRollup -> readUtf8Text rollupPath
-      | hasSmaps -> readUtf8Text smapPath
-      | otherwise -> pure Text.empty
+    | hasRollup -> readUtf8Text rollupPath
+    | hasSmaps -> readUtf8Text smapPath
+    | otherwise -> pure Text.empty
 
 
 overallTotals :: [MemUsage] -> (Int, Int)
@@ -587,3 +602,35 @@ withPid (pid, name, pp) = ((pid, name), pp)
 -}
 dropId :: Indexer ProcName
 dropId (_pid, name, pp) = (name, pp)
+
+
+byPrintOrder ::
+  (Ord c) =>
+  (((c, MemUsage) -> Int) -> (c, MemUsage) -> (c, MemUsage) -> Ordering) ->
+  PrintOrder ->
+  (c, MemUsage) ->
+  (c, MemUsage) ->
+  Ordering
+byPrintOrder f Swap = f $ muSwap . snd
+byPrintOrder f Shared = f $ muShared . snd
+byPrintOrder f Private = f $ muPrivate . snd
+byPrintOrder f Count = f $ muCount . snd
+
+
+byPrintOrder' ::
+  (Ord a) =>
+  Bool ->
+  Maybe PrintOrder ->
+  (a, MemUsage) ->
+  (a, MemUsage) ->
+  Ordering
+byPrintOrder' reversed mbOrder =
+  let cmpUsage = if reversed then comparing else comparing'
+      cmpName = if reversed then comparing else comparing'
+      byName = cmpName fst
+      byUsage = byPrintOrder cmpUsage
+   in maybe byName byUsage mbOrder
+
+
+comparing' :: (Ord a) => (b -> a) -> b -> b -> Ordering
+comparing' f a b = compare (Down $ f a) (Down $ f b)