+++ src/System/MemInfo/Choices.hs
@@ -1,7 +1,6 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE GeneralisedNewtypeDeriving #-}
 
 {- |
 Module      : System.MemInfo.Choices
@@ -14,6 +13,7 @@ __printmem__ command
 -}
 module System.MemInfo.Choices (
   Choices (..),
+  PrintOrder (..),
   cmdInfo,
   getChoices,
 ) where
@@ -51,8 +51,10 @@ data Choices = Choices
   , choiceOnlyTotal :: !Bool
   , choiceByPid :: !Bool
   , choiceShowSwap :: !Bool
+  , choiceReversed :: !Bool
   , choiceWatchSecs :: !(Maybe Natural)
   , choicePidsToShow :: !(Maybe (NonEmpty ProcessID))
+  , choicePrintOrder :: !(Maybe PrintOrder)
   }
   deriving (Eq, Show, Generic)
 
@@ -69,59 +71,69 @@ parseChoices =
     <*> parseOnlyTotal
     <*> parseDiscriminateByPid
     <*> parseShowSwap
+    <*> parseReversed
     <*> optional parseWatchPeriodSecs
     <*> optional parseChoicesPidsToShow
+    <*> optional parsePrintOrder
 
 
 parseChoicesPidsToShow :: Parser (NonEmpty ProcessID)
 parseChoicesPidsToShow =
-  some1 $
-    option positiveNum $
-      short 'p'
-        <> long "pids"
-        <> metavar "<pid1> [ -p pid2 ... -p pidN ]"
-        <> help "Only show memory usage of the specified PIDs"
+  some1
+    $ option positiveNum
+    $ short 'p'
+    <> long "pids"
+    <> metavar "<pid1> [ -p pid2 ... -p pidN ]"
+    <> help "Only show memory usage of the specified PIDs"
 
 
 parseSplitArgs :: Parser Bool
 parseSplitArgs =
-  switch $
-    short 's'
-      <> long "split-args"
-      <> help "Show and separate by all command line arguments"
+  switch
+    $ short 's'
+    <> long "split-args"
+    <> help "Show and separate by all command line arguments"
 
 
 parseOnlyTotal :: Parser Bool
 parseOnlyTotal =
-  switch $
-    short 't'
-      <> long "total"
-      <> help "Only show the total value"
+  switch
+    $ short 't'
+    <> long "total"
+    <> help "Only show the total value"
+
+
+parseReversed :: Parser Bool
+parseReversed =
+  switch
+    $ short 'r'
+    <> long "reverse"
+    <> help "Reverses the output order so that output descends on the sorting field"
 
 
 parseDiscriminateByPid :: Parser Bool
 parseDiscriminateByPid =
-  switch $
-    short 'd'
-      <> long "discriminate-by-pid"
-      <> help "Show by process rather than by program"
+  switch
+    $ short 'd'
+    <> long "discriminate-by-pid"
+    <> help "Show by process rather than by program"
 
 
 parseShowSwap :: Parser Bool
 parseShowSwap =
-  switch $
-    short 'S'
-      <> long "show_swap"
-      <> help "Show swap information"
+  switch
+    $ short 'S'
+    <> long "show_swap"
+    <> help "Show swap information"
 
 
 parseWatchPeriodSecs :: Parser Natural
 parseWatchPeriodSecs =
-  option positiveNum $
-    short 'w'
-      <> long "watch"
-      <> metavar "N"
-      <> help "Measure and show memory every N seconds (N > 0)"
+  option positiveNum
+    $ short 'w'
+    <> long "watch"
+    <> metavar "N"
+    <> help "Measure and show memory every N seconds (N > 0)"
 
 
 positiveNum :: (Read a, Ord a, Num a) => ReadM a
@@ -132,3 +144,21 @@ positiveNum =
       | otherwise = readerError "Value must be greater than 0"
    in
     auto >>= checkPositive
+
+
+parsePrintOrder :: Parser PrintOrder
+parsePrintOrder =
+  option auto
+    $ short 'b'
+    <> long "order-by"
+    <> metavar "<Private | Swap | Shared | Count>"
+    <> help "Orders the output by ascending values of the given field"
+
+
+-- | Determines the order in which @MemUsages@ in a report are printed out
+data PrintOrder
+  = Swap
+  | Private
+  | Shared
+  | Count
+  deriving (Eq, Show, Read, Generic)