+++ ChangeLog.md
@@ -2,6 +2,13 @@
 
 `mem-info` uses [PVP Versioning][1].
 
+## 0.2.0.0 -- 2024-01-28
+
+- Simplify the output when the -d (--discriminate-by-pid) flag is used
+
+- Add options -b (--order-by) and -r (--reverse) to change the ordering of the
+  output
+
 ## 0.1.0.1 -- 2024-01-17
 
 - Adjusted dependency bounds to remove any stale dependencies